+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from './features/header/Header';
+import Header from './features/Header/Header';
 
 class App extends Component {
   render() {
