+++ src/features/cards/CardList/CardList.test.tsx
@@ -2,7 +2,7 @@ import { screen } from '@testing-library/dom';
 import { render } from '@testing-library/react';
 import { describe, test } from 'vitest';
 import { CARDS } from '../../../mocks/cards.mock';
-import CardList from './Card-list';
+import CardList from './CardList';
 
 describe('CardList', () => {
   test('Should have cards', () => {
